Only really added Voidmetal Drum to this Release, still feeling bad...

+++ src/main/java/gregapi/GT_API_Proxy.java
@@ -1074,7 +1074,7 @@ public abstract class GT_API_Proxy extends Abstract_Proxy implements IGuiHandler
 		long rFuelValue = 0;
 		if (aFuel.getItem() instanceof MultiItemRandom) {
 			Short tFuelValue = ((MultiItemRandom)aFuel.getItem()).mBurnValues.get(ST.meta_(aFuel));
-			if (tFuelValue != null) rFuelValue = (short)Math.max(rFuelValue, tFuelValue);
+			if (tFuelValue != null) rFuelValue = Math.max(rFuelValue, tFuelValue);
 		} else {
 			if (OM.is_(OD.logWood, aFuel)) return TICKS_PER_SMELT * 6;
 			if (IL.IC2_Resin.equal(aFuel, F, T)) return TICKS_PER_SMELT / 2;
